requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.1.1
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
celery==5.2.6
kombu==5.2.4
